+++ src/components/App/Header/InboxPage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import InboxContent from './InboxContent';
+
+const InboxPage = () => {
+  return <InboxContent messages={[]} />;
+};
+
+export default InboxPage;
