--- src/components/App/Header/MailboxPage.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import MailboxContent from './MailboxContent';
-
-const MailboxPage = () => {
-  return <MailboxContent messages={[]} />;
-};
-
-export default MailboxPage;
